+++ crates/controller/src/api/v1/mod.rs
@@ -8,7 +8,6 @@
 //! - `/rooms/{room_id}/start_invited` ([POST](rooms::start_invited))
 //! - `/rooms/{room_id}/invites ([GET](invites::get_invites), [POST](invites::add_invite))
 //! - `/rooms/{room_id}/invites/{invite_code} ([GET](invites::get_invite), [PUT](invites::update_invite), [DELETE](invites::delete_invite)])
-//! - `/rooms/sip/start ([POST](rooms::sip_start))
 //! - `/rooms/{room_id}/sip ([GET](sip_configs::get), [PUT](sip_configs::put), [DELETE](sip_configs::delete))
 //! - `/rooms/{room_id}/legal_votes ([GET](legal_vote::get_all_for_room))
 //! - `/turn` ([GET](turn::get))
@@ -17,6 +16,7 @@
 //! - `/users/find` ([GET](users::find))
 //! - `/legal_votes` ([GET](legal_vote::get_all))
 //! - `/legal_votes/{legal_vote_id}` ([GET](legal_vote::get_specific))
+//! - `/services/call_in/start ([POST](services::call_in::start))
 
 pub use request::{CursorPaginationQuery, PagePaginationQuery};
 pub use response::{ApiResponse, DefaultApiResult};
@@ -31,6 +31,7 @@ pub mod middleware;
 mod request;
 pub mod response;
 pub mod rooms;
+pub mod services;
 pub mod sip_configs;
 pub mod turn;
 pub mod users;